get read io tests working
